Move header component to the ui folder
