+++ site/gatsby-site/src/components/ui/Header.js
@@ -7,9 +7,9 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faBars, faRssSquare } from '@fortawesome/free-solid-svg-icons';
 import { faTwitterSquare } from '@fortawesome/free-brands-svg-icons';
 
-import Link from './ui/Link';
-import LoadingProvider from './mdxComponents/loading';
-import config from '../../config.js';
+import Link from './Link';
+import LoadingProvider from '../mdxComponents/loading';
+import config from '../../../config.js';
 
 const help = require('./images/help.svg');
 
@@ -25,10 +25,10 @@ if (isSearchEnabled && config.header.search.indexName) {
   });
 }
 
-import Sidebar from './sidebar';
+import Sidebar from '../sidebar';
 
 const LoadableComponent = Loadable({
-  loader: () => import('./search/index'),
+  loader: () => import('../search/index'),
   loading: LoadingProvider,
 });
